Refactor aave handlers to extract reusable code
